autonomous_growth/requirements.txt
aiohttp==3.9.1
schedule==1.2.1
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
asyncio==3.4.3
psutil==5.9.6
aiofiles==23.2.1
cryptography==41.0.7
web3==6.11.3
python-binance==1.0.19
tweepy==4.14.0
google-api-python-client==2.108.0
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
